Full Part C 4 Param search + adjustments to corner plot & switch to float 32 from float64 in loc_types.py

+++ microlensing/loc_types.py
@@ -2,4 +2,4 @@
 import numpy as np
 import numpy.typing as npt
 
-NDFloatArray = npt.NDArray[np.float64]
+NDFloatArray = npt.NDArray[np.float32]
